+++ socks.json
@@ -1,0 +1,4 @@
+{
+  "Socks{localDate=2023-02-21, size=S36_37, color=RED, cottonPart=60}": 20,
+  "Socks{localDate=2023-02-21, size=S36_37, color=RED, cottonPart=90}": 270
+}
