+++ src/main/java/sky/pro/socksappwebstore/model/Socks.java
@@ -8,7 +8,6 @@ import java.time.LocalDate;
 import java.util.Objects;
 
 @Data
-@NoArgsConstructor
 public class Socks {
     private LocalDate localDate;
     private Size size;
