+++ src/main/java/sky/pro/socksappwebstore/services/SocksService.java
@@ -2,10 +2,12 @@ package sky.pro.socksappwebstore.services;
 
 import sky.pro.socksappwebstore.model.Color;
 import sky.pro.socksappwebstore.model.Size;
+import sky.pro.socksappwebstore.model.Socks;
 import sky.pro.socksappwebstore.model.SocksBatch;
 
 import java.io.IOException;
 import java.nio.file.Path;
+import java.util.Map;
 
 public interface SocksService {
     void accept(SocksBatch socksBatch);
@@ -16,4 +18,10 @@ public interface SocksService {
 
     int getCount(Color color, Size size, int cottonMin, int cottonMax);
     Path createAllSocks() throws IOException;
+
+    void save(SocksBatch socksBatch);
+
+    int remove(SocksBatch socksBatch);
+
+    Map<Socks, Integer> getAll();
 }
