+++ src/main/java/sky/pro/socksappwebstore/services/impl/SocksServiceImpl.java
@@ -3,18 +3,17 @@ package sky.pro.socksappwebstore.services.impl;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import sky.pro.socksappwebstore.exception.ValidationException;
 import sky.pro.socksappwebstore.model.Color;
 import sky.pro.socksappwebstore.model.Size;
 import sky.pro.socksappwebstore.model.Socks;
 import sky.pro.socksappwebstore.model.SocksBatch;
-import sky.pro.socksappwebstore.services.FileSocksService;
 import sky.pro.socksappwebstore.services.FilesService;
 import sky.pro.socksappwebstore.services.SocksService;
 import sky.pro.socksappwebstore.services.ValidationService;
 
+import javax.annotation.PostConstruct;
 import java.io.IOException;
 import java.io.Writer;
 import java.nio.file.Files;
@@ -24,29 +23,39 @@ import java.util.HashMap;
 import java.util.Map;
 
 @Service
-@AllArgsConstructor
 public class SocksServiceImpl implements SocksService {
-    private HashMap<Socks, Integer> socksMap = new HashMap<>();
-    private final FileSocksService fileSocksService;
+    private static Map<Socks, Integer> socksMap = new HashMap<>();
     private final ValidationService validationService;
     private final FilesService filesService;
 
+    public SocksServiceImpl(ValidationService validationService, FilesService filesService) {
+        this.validationService = validationService;
+        this.filesService = filesService;
+    }
+//    @PostConstruct
+//    private void bim() {
+//        try {
+//            readFromFile();
+//        } catch (Exception e) {
+//            e.printStackTrace();
+//        }
+//    }
     @Override
     public void accept(SocksBatch socksBatch) {
         validationService.validate(socksBatch);
-        fileSocksService.save(socksBatch);
+        save(socksBatch);
     }
 
     @Override
     public int issuence(SocksBatch socksBatch) {
         validationService.validate(socksBatch);
-        return fileSocksService.remove(socksBatch);
+        return remove(socksBatch);
     }
 
     @Override
     public int reject(SocksBatch socksBatch) {
         validationService.validate(socksBatch);
-        return fileSocksService.remove(socksBatch);
+        return remove(socksBatch);
     }
 
     @Override
@@ -54,7 +63,7 @@ public class SocksServiceImpl implements SocksService {
         if (validationService.validate(color, size, cottonMin, cottonMax)) {
             throw new ValidationException("Ошибка валидации параметров: цвет, размер, содержание хлопка в носках");
         }
-        HashMap<Socks, Integer> socksMap = fileSocksService.getAll();
+        Map<Socks, Integer> socksMap = getAll();
         for (Map.Entry<Socks, Integer> element : socksMap.entrySet()) {
             Socks socks = element.getKey();
             if (socks.getColor().equals(color)
@@ -70,7 +79,7 @@ public class SocksServiceImpl implements SocksService {
 
     @Override
     public Path createAllSocks() throws IOException {
-        Path path = filesService.getAllsocksFile().toPath();
+        Path path = filesService.getSocksFile().toPath();
         String listStop = "*";
         for (Map.Entry<Socks, Integer> element : socksMap.entrySet()) {
             try (Writer writer = Files.newBufferedWriter(path, StandardOpenOption.APPEND)) {
@@ -88,21 +97,58 @@ public class SocksServiceImpl implements SocksService {
         return path;
     }
 
+
+
+    @Override
+    public void save(SocksBatch socksBatch) {
+        Socks socks = socksBatch.getSocks();
+        if (socksMap.containsKey(socks)) {
+            socksMap.replace(socks, socksMap.get(socks) + socksBatch.getQuantity());
+            saveToFile();
+        } else {
+            socksMap.put(socks, socksBatch.getQuantity());
+            saveToFile();
+        }
+
+    }
+
+    @Override
+    public int remove(SocksBatch socksBatch) {
+        Socks socks = socksBatch.getSocks();
+        if (socksMap.containsKey(socks)) {
+            int totalQuantity = socksMap.get(socks);
+            if (totalQuantity > socksBatch.getQuantity()) {
+                socksMap.replace(socks, socksMap.get(socks) - socksBatch.getQuantity());
+                saveToFile();
+                return socksBatch.getQuantity();
+            } else {
+                socksMap.remove(socks);
+                saveToFile();
+                return totalQuantity;
+            }
+        }
+        return 0;
+    }
+
+    public Map<Socks, Integer> getAll() {
+        return socksMap;
+    }
+
+    public void saveToFile() {
+        try {
+            String json = new ObjectMapper().writeValueAsString(socksMap);
+            filesService.saveToFile(json);
+        } catch (JsonProcessingException e) {
+            throw new RuntimeException(e);
+        }
+    }
     public void readFromFile() {
         try {
             String json = filesService.readFromFile();
-            socksMap = new ObjectMapper().readValue(json, new TypeReference<HashMap<Socks, Integer>>() {
+            socksMap = new ObjectMapper().readValue(json, new TypeReference<Map<Socks, Integer>>() {
             });
         } catch (JsonProcessingException e) {
             throw new RuntimeException(e);
         }
     }
-//    @PostConstruct
-//    private void bim() {
-//        try {
-//            readFromFile();
-//        } catch (Exception e) {
-//            e.printStackTrace();
-//        }
-//    }
 }
