+++ src/main/resources/application.properties
@@ -1,4 +1,2 @@
 path.to.socks.file=src/main/resources
 name.of.socks.file=socks.json
-path.to.allsocks.file=src/main/resources
-name.of.allsocks.file=allsocks.json
