--- src/main/resources/socks.json
@@ -1,4 +1,0 @@
-{
-  "Socks{localDate=2023-02-21, size=S36_37, color=RED, cottonPart=90}": 880,
-  "Socks{localDate=2023-02-21, size=S36_37, color=RED, cottonPart=70}": 80
-}
